% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult}
\alias{adult}
\title{adult—A widely used data set to predict if a person makes over $50,000/year. Automatic solution is possible using logistic analysis in the Ensembles package}
\format{
## `Adult`
A data set with 48842 rows and 14 columns

\describe{
 \item{age}{a person's age in years}
 \item{workclass}{categorial, based on income. Values include: Private, Self-emp-not-inc, Self-emp-inc, Federal-gov, Local-gov, State-gov, Without-pay, Never-worked.}
 \item{fnlwgt}{integer used in the calculations}
 \item{education}{categorical, lists education level. Values include: Bachelors, Some-college, 11th, HS-grad, Prof-school, Assoc-acdm, Assoc-voc, 9th, 7th-8th, 12th, Masters, 1st-4th, 10th, Doctorate, 5th-6th, Preschool.}
 \item{education-num}{integer, gives the numerical value of a person's education}
 \item{maritalstatus}{categorical, values include: Married-civ-spouse, Divorced, Never-married, Separated, Widowed, Married-spouse-absent, Married-AF-spouse.}
 \item{occupation}{categorical, values include: Tech-support, Craft-repair, Other-service, Sales, Exec-managerial, Prof-specialty, Handlers-cleaners, Machine-op-inspct, Adm-clerical, Farming-fishing, Transport-moving, Priv-house-serv, Protective-serv, Armed-Forces.}
 \item{relationship}{categorical, values include: Wife, Own-child, Husband, Not-in-family, Other-relative, Unmarried}
 \item{race}{categorical, values include: White, Asian-Pac-Islander, Amer-Indian-Eskimo, Other, Black}
 \item{sex}{binary, values include female, male}
 \item{capital-gain}{integer}
 \item{capital-loss}{integer}
 \item{hours-per-week}{integer}
 \item{native-country}{categorical, values include: United-States, Cambodia, England, Puerto-Rico, Canada, Germany, Outlying-US(Guam-USVI-etc), India, Japan, Greece, South, China, Cuba, Iran, Honduras, Philippines, Italy, Poland, Jamaica}
 \item{native-country}{categorical, values include: Vietnam, Mexico, Portugal, Ireland, France, Dominican-Republic, Laos, Ecuador, Taiwan, Haiti, Columbia, Hungary, Guatemala, Nicaragua, Scotland, Thailand, Yugoslavia, El-Salvador, Trinadad&Tobago, Peru, Hong, Holand-Netherlands}
 \item{income}{binary, <50k, >=50k}

}
}
\source{
<https://archive.ics.uci.edu/dataset/2/adult>
}
\usage{
adult
}
\description{
A very widely used data set to predict if a person earns over $50,000 per year. This is also known as the "Census Income" data set.
It is multivariate, using both categorical and integer values.
Additional infomation provided from the source:
"Additional Information
Extraction was done by Barry Becker from the 1994 Census database.  A set of reasonably clean records was extracted using the following conditions: ((AAGE>16) && (AGI>100) && (AFNLWGT>1)&& (HRSWK>0))
Prediction task is to determine whether a person makes over 50K a year.
}
\keyword{datasets}
