% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank_marketing}
\alias{bank_marketing}
\title{bank_marketing—A data set related to direct marketing calls by a Portuguese banking institution. Automatic solution is possible as a classification problem in the Ensembles package.}
\format{
An object of class \code{data.frame} with 45211 rows and 17 columns.
}
\source{
<https://archive.ics.uci.edu/dataset/222/bank+marketing>
}
\usage{
bank_marketing
}
\description{
From the source: This is a data set that is related to direct marketing calls by a Portuguese banking institution. The classification goal is to predict if
the client will subscribe a short term deposit (variable y)

\describe{
 \item{age}{a person's age}
 \item{job}{categorical, a person's job classification. Values: categorical: 'admin.','blue-collar','entrepreneur','housemaid','management','retired','self-employed','services','student','technician','unemployed','unknown'}
 \item{marital}{cagegorical, values include: 'divorced','married','single','unknown'; note: 'divorced' means divorced or widowed}
 \item{education}{categorical, values include: 'basic.4y','basic.6y','basic.9y','high.school','illiterate','professional.course','university.degree','unknown'}
 \item{default}{binary, is credit in default?}
 \item{balance}{average yearly balance}
 \item{housing}{has housing loan?}
 \item{loan}{has personal loan?}
 \item{contact}{method of contact, values include: 'cellular','telephone'}
 \item{day_of_week}{date, last contact date of week}
 \item{month}{last contact month of year, values include: 'jan', 'feb', 'mar', ..., 'nov', 'dec'}
 \item{campaign}{integer, the number of contacts performed during this campaign}
 \item{pdays}{integer, number of days that passed by after the client was last contacted from a previous campaign (numeric; -1 means client was not previously contacted)}
 \item{previous}{integer, number of contacts performed before this campaign and for this client}
 \item{poutcome}{categorical, values include: 'failure','nonexistent','success'}
 \item{outcome}{binary, has the client subscribed a term deposit}
 }
}
\keyword{datasets}
