% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{classification—function to perform classification analysis and return results to the user.}
\usage{
classification(
  data,
  colnum,
  numresamples,
  do_you_have_new_data = c("Y", "N"),
  how_to_handle_strings = c(0("No strings"), 1("Strings as factors")),
  save_all_trained_models = c("Y", "N"),
  use_parallel = c("Y", "N"),
  train_amount,
  test_amount,
  validation_amount
)
}
\arguments{
\item{data}{a data set that includes classification data. For example, the Carseats data in the ISLR package}

\item{colnum}{the number of the column. For example, in the Carseats data this is column 7, ShelveLoc with three values, Good, Medium and Bad}

\item{numresamples}{the number of times to resample the analysis}

\item{do_you_have_new_data}{asks if the user has new data to be analyzed using the trained models that were just developed}

\item{how_to_handle_strings}{Converts strings to factor levels}

\item{save_all_trained_models}{Gives the user the option to save all trained models in the Environment}

\item{use_parallel}{"Y" or "N" for parallel processing}

\item{train_amount}{set the amout for the training data}

\item{test_amount}{set the amount for the testing data}

\item{validation_amount}{Set the amount for the validation data}
}
\value{
a full analysis, including data visualizations, statistical summaries, and a full report on the results of 35 models on the data
}
\description{
classification—function to perform classification analysis and return results to the user.
}
\examples{
classification(data = dry_beans_small,
                colnum = 17,
                numresamples = 2,
                do_you_have_new_data = "N",
                how_to_handle_strings = 0,
                save_all_trained_models = "N",
                use_parallel = "Y",
                train_amount = 0.60,
                test_amount = 0.20,
                validation_amount = 0.20
                )


}
