% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Concrete_Data}
\alias{Concrete_Data}
\title{Concrete_Data—A data set to measure the compressive strength of concrete. Automatic solution as a numerical problem in the Ensembles package.}
\format{
An object of class \code{data.frame} with 1030 rows and 9 columns.
}
\source{
<https://archive.ics.uci.edu/dataset/165/concrete+compressive+strength>
}
\usage{
Concrete_Data
}
\description{
This is sometimes called the concrete compressive strength data set.

\describe{
 \item{Cement}{quantitative, kg in a m^3 mixture}
 \item{Blast_Furnace_Slag}{quantitative,  kg in a m^3 mixtur}
 \item{Fly_Ash}{quantitative, kg in a m^3 mixture}
 \item{Water}{quantitative, kg in a m^3 mixture}
 \item{Superplasticizer}{quantitative, kg in a m^3 mixture}
 \item{Coarse_Aggregate}{quantitative, kg in a m^3 mixture}
 \item{Fine_Aggregate}{quantitative, kg in a m^3 mixture}
 \item{Age}{quantitative, Day, 1~365}
 \item{Strength}{quantitative, measured in MPa}
}
}
\keyword{datasets}
