% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dry_beans}
\alias{dry_beans}
\title{dry_beans—A classification problem of 7 different types of dry beans. Automatic solution as a classification problem using the Ensembles package.}
\format{
An object of class \code{data.frame} with 13611 rows and 17 columns.
}
\source{
<https://www.kaggle.com/datasets/muratkokludataset/dry-bean-dataset>
}
\usage{
dry_beans
}
\description{
Abstract: Images of 13,611 grains of 7 different registered dry beans were taken with a high-resolution camera. A total of 16 features; 12 dimensions and 4 shape forms, were obtained from the grains.
Citation Requests / Acknowledgements:
KOKLU, M. and OZKAN, I.A., (2020), Multiclass Classification of Dry Beans Using Computer Vision and Machine Learning Techniques. Computers and Electronics in Agriculture, 174, 105507.
DOI: https://doi.org/10.1016/j.compag.2020.105507
Relevant Information:
Seven different types of dry beans were used in this research, taking into account the features such as form, shape, type, and structure by the market situation. A computer vision system was developed to distinguish seven different registered varieties of dry beans with similar features in order to obtain uniform seed classification. For the classification model, images of 13,611 grains of 7 different registered dry beans were taken with a high-resolution camera. Bean images obtained by computer vision system were subjected to segmentation and feature extraction stages, and a total of 16 features; 12 dimensions and 4 shape forms, were obtained from the grains.

\describe{
\item{Area}{The area of a bean zone and the number of pixels within its boundaries.}
\item{Permieter}{Bean circumference is defined as the length of its border.}
\item{MajorAxisLength}{The distance between the ends of the longest line that can be drawn from a bean.}
\item{MinorAxisLength}{The longest line that can be drawn from the bean while standing perpendicular to the main axis.}
\item{AspectRation}{Defines the relationship between MajorAxisLength and MinorAxisLength}
\item{Eccentricity}{Eccentricity of the ellipse having the same moments as the region}
\item{ConvexArea}{Number of pixels in the smallest convex polygon that can contain the area of a bean seed}
\item{EquivDiameter}{The diameter of a circle having the same area as a bean seed area}
\item{Extent}{The ratio of the pixels in the bounding box to the bean area}
\item{Solidity}{Also known as convexity. The ratio of the pixels in the convex shell to those found in beans}
\item{roundness}{Calculated with the following formula: (4piA)/(P^2)}
\item{Compactness}{Measures the roundness of an object: EquivalentDiamter/MajorAxisLength}
\item{ShapeFactor1}{Shape Factor}
\item{ShapeFactor2}{Shape Factor}
\item{ShapeFactor3}{Shape Factor}
\item{ShapeFactor4}{Shape Factor}
\item{Class}{One of Seker, Barbunya, Bombay, Cali, Dermosan, Horoz and Sira}

}
}
\keyword{datasets}
