% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dry_beans_small}
\alias{dry_beans_small}
\title{dry_beans_small-A small version of the entire dry beans data set}
\format{
An object of class \code{data.frame} with 700 rows and 17 columns.
}
\source{
: https://archive.ics.uci.edu/dataset/602/dry+bean+dataset
}
\usage{
dry_beans_small
}
\description{
This is a famous data set used for classification. This example is a stratified sample of the full data set.

\describe{
\item{Area}{The area of a bean zone and the number of pixels within its boundaries}
\item{Perimeter}{Bean circumference is defined as the length of its border}
\item{MajorAxisLength}{The distance between the ends of the longest line that can be drawn from a bean}
\item{MinorAxisLength}{The longest line that can be drawn from the bean while standing perpendicular to the main axis}
\item{AspectRatio}{Defines the relationship between MajorAxisLength and MinorAxisLength}
\item{Eccentricity}{Eccentricity of the ellipse having the same moments as the region}
\item{ConvexArea}{Number of pixels in the smallest convex polygon that can contain the area of a bean seed}
\item{EquivDiameter}{Equivalent diameter: The diameter of a circle having the same area as a bean seed area}
\item{Extent}{The ratio of the pixels in the bounding box to the bean area}
\item{Solidity}{Also known as convexity. The ratio of the pixels in the convex shell to those found in beans}
\item{Roundness}{Calculated with the following formula: (4piA)/(P^2)}
\item{Compactness}{Measures the roundness of an object}
\item{ShapeFactor1}{ShapeFactor1}
\item{ShapeFactor2}{ShapeFactor2}
\item{ShapeFactor3}{ShapeFactor3}
\item{ShapeFactor4}{ShapeFactor4}
\item{Class}{(Seker, Barbunya, Bombay, Cali, Dermosan, Horoz and Sira)}
}
}
\keyword{datasets}
