% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecasting}
\alias{forecasting}
\title{forecasting—function to perform time series analysis and return the results to the user.}
\usage{
forecasting(
  time_series_data,
  train_amount,
  number,
  time_interval = c("Q", "M", "W"),
  use_parallel = c("Y", "N")
)
}
\arguments{
\item{time_series_data}{a time series}

\item{train_amount}{The amount for the training set. For example, 0.60 is 60 percent.}

\item{number}{Number of time intervals to forecast. For example, 3.}

\item{time_interval}{user states whether the time interval is quarterly ("Q"), monthly ("M"), or weekly ("W").}

\item{use_parallel}{"Y" or "N" for parallel processing}
}
\value{
A series of summary reports and visualizations to fully describe the time series: Forecast accuracy, forecast numbers, forecast plot, innovation residuals,

best autocorrelation function (ACF), plot of best histogram of residuals, plot of best actual vs predicted, plot of best actual vs trend

plot of best actual vs seasonally adjusted
}
\description{
forecasting—function to perform time series analysis and return the results to the user.
}
\examples{
data <- all_employees (note this example will take a few minutes to run, the function does give updates of which model is build built)
forecasting(time_series_data = total_nonfarm,
             train_amount = 0.60,
             number = 3,
             time_interval = "M",
             use_parallel = "Y") # note this example takes several minutes to run on a normal computer.
}
