% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Monthly_Labor}
\alias{Monthly_Labor}
\title{Monthly_Labor-This is a set of monthly time series reports from the United States Bureau of Labor Statistics. It lists the number of employees in each of the following areas:
The current month, All_Employees, Construction, Education, Finance, Government, Information, Hospitality, Manufacturing, Oil_and_Gas, Retail, Temporary, and the Unemployment rate.}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 216 rows and 14 columns.
}
\source{
<https://data.bls.gov/>
}
\usage{
Monthly_Labor
}
\description{
These are 13 time series reports from the United States Bureau of Labor Statistics, all in one document.

\describe{
The original data is from the United States Department of Labor Statistics. The data consists of 14 columns: A Month column, 12 sections of the United States economy (such as Construction or Education),
and the total unemployment rate for the nation.

\item{Month}{The month being reported, starting with January 2006}
\item{All_Employees}{The total number of non-farm employees for a specific month, series number: CES0000000001}
\item{Construction}{The number of people working in Construction for that month, series number: CES2000000001}
\item{Education}{The number of people working in Education for that month, series number: CES6500000001}
\item{Finance}{The number of people working in Finance for that month, series number: CES5500000001}
\item{Government}{The number of people working in government for that month, series number: CES9000000001}
\item{Information}{The number of people working in Information for that month, series number: CES5000000001}
\item{Hospitality}{The number of people working in Hospitality for that month, series number: CES7000000001}
\item{Manufacturing}{The number of people working in Manufacturing for that month, series number: CES3000000001}
\item{Oil_and_Gas}{The number of people working in Oil and Gas for that month, series number: CES1021100001}
\item{Prof_Svcs}{The number of people working in Professional Services for that month, series number: CES6000000001}
\item{Tempoary}{The number of people working in the tempoary staffing industry that month, series number: CES6056132001}
\item{Unemployment}{The national (United States) unemployment rate for that month, LNS14000000}
}
}
\keyword{datasets}
